Trying to reverse board
I am encountering some problems in the chessboard inversion feature; with the F key command it is possible to change a variable to be
passed later to the various functions, but it is not sufficient to pass it only to the function that shows the pieces, since both
the selection and the generation of moves require the updated position of the pieces. To be reviewed.